CortexM: moved creation of default Cortex-M memory map to CoreSightTarget.
With the default map only being present on the CortexM instance, it
had no effect for loading to RAM.
